requirements.txt
torch==1.8.1
numpy==1.18.1
pytorch-lightning==1.2.7
scikit-learn==0.21.3
tqdm==4.43.0
hydra-core==1.0.6
transformers[sentencepiece]==4.5.1
Jinja2==2.11.1
